# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /masters/chargers:
    get:
      operationId: listChargerMasters
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/StatusParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: List of charger masters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/MasterListResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ChargerMaster"
    post:
      operationId: createChargerMaster
      tags: [masters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargerMasterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerMaster"
  /masters/chargers/{id}:
    get:
      operationId: getChargerMaster
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Single charger master
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerMaster"
    patch:
      operationId: updateChargerMaster
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargerMasterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerMaster"
  /masters/chargers/{id}/status:
    patch:
      operationId: toggleChargerMasterStatus
      tags: [masters]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterStatusToggle"
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerMaster"

  # Test Equipment Masters
...
  /manufacturing/charger-units:
    get:
      operationId: listChargerUnits
      tags: [manufacturing]
      summary: List operational charger units
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [available, busy, maintenance] }
        - name: pageSize
          in: query
          schema: { type: integer, default: 50 }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: Charger units list
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChargerUnit"
                  total: { type: integer }
    post:
      operationId: createChargerUnit
      tags: [manufacturing]
      summary: Register a new charger unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargerUnitInput"
      responses:
        "201":
          description: Charger unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerUnit"
...
  /manufacturing/charger-units/{id}:
    get:
      operationId: getChargerUnit
      tags: [manufacturing]
      summary: Get a single charger unit
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Charger unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerUnit"
    patch:
      operationId: updateChargerUnit
      tags: [manufacturing]
      summary: Update charger unit details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargerUnitUpdate"
      responses:
        "200":
          description: Updated charger unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerUnit"
  /manufacturing/charger-units/{id}/status:
    patch:
      operationId: updateChargerUnitStatus
      tags: [manufacturing]
      summary: Change a charger unit's operational status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargerUnitStatusInput"
      responses:
        "200":
          description: Updated charger unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargerUnit"

  # ── Manufacturing — Timeline ──────────────────────────────────────────────
...
  /manufacturing/orders/{id}/qc-approval:
    get:
      operationId: getQcApproval
      tags: [manufacturing]
      summary: Get QC approval for an order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: QC approval record or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcApproval"
    post:
      operationId: createQcApproval
      tags: [manufacturing]
      summary: Submit QC approval or rejection (Plant Manager only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcApprovalInput"
      responses:
        "201":
          description: QC approval created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcApproval"

  # ── Manufacturing — Rework Queue ─────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SearchParam:
      name: search
      in: query
      required: false
      schema:
        type: string
    StatusParam:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, inactive]
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        default: 25
  schemas:
...
    MasterStatus:
      type: string
      enum: [active, inactive]
    MasterStatusToggle:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/MasterStatus"
    Attachment:
      type: object
      required: [id, name, url, fileType, uploadedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        url: { type: string }
        fileType: { type: string }
        uploadedAt: { type: string, format: date-time }
    MasterCommon:
      type: object
      required: [id, code, name, status, revision_number, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/MasterStatus" }
        revision_number: { type: integer }
        effective_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        created_by: { type: ["string", "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_by: { type: ["string", "null"], format: uuid }
        updated_at: { type: string, format: date-time }
    MasterCommonInput:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        effective_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    MasterCommonUpdate:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        effective_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    MasterListMeta:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
    MasterListResponse:
      type: object
      required: [items, meta]
      properties:
        meta: { $ref: "#/components/schemas/MasterListMeta" }

    # Product Master
...
    ChargerMaster:
      allOf:
        - $ref: "#/components/schemas/MasterCommon"
        - type: object
          properties:
            manufacturer: { type: string }
            model: { type: string }
            output_voltage_v: { type: number }
            output_current_a: { type: number }
            power_kw: { type: number }
            communication_protocol: { type: string }
            datasheet_url: { type: ["string", "null"] }
    ChargerMasterInput:
      allOf:
        - $ref: "#/components/schemas/MasterCommonInput"
        - type: object
          required: [manufacturer, model, output_voltage_v, output_current_a, power_kw, communication_protocol]
          properties:
            manufacturer: { type: string }
            model: { type: string }
            output_voltage_v: { type: number }
            output_current_a: { type: number }
            power_kw: { type: number }
            communication_protocol: { type: string }
            datasheet_url: { type: ["string", "null"] }
    ChargerMasterUpdate:
      allOf:
        - $ref: "#/components/schemas/MasterCommonUpdate"
        - type: object
          properties:
            manufacturer: { type: string }
            model: { type: string }
            output_voltage_v: { type: number }
            output_current_a: { type: number }
            power_kw: { type: number }
            communication_protocol: { type: string }
            datasheet_url: { type: ["string", "null"] }

    # Test Equipment Master
...
    ChargerUnit:
      type: object
      required: [id, chargerCode, model, manufacturer, serialNumber, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        chargerCode: { type: string }
        model: { type: string }
        manufacturer: { type: string }
        serialNumber: { type: string }
        outputVoltageV: { type: ["number", "null"] }
        maxCurrentA: { type: ["number", "null"] }
        status: { type: string, enum: [available, busy, maintenance] }
        currentOrderId: { type: ["string", "null"], format: uuid }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ChargerUnitInput:
      type: object
      required: [chargerCode, model, manufacturer, serialNumber]
      properties:
        chargerCode: { type: string }
        model: { type: string }
        manufacturer: { type: string }
        serialNumber: { type: string }
        outputVoltageV: { type: ["number", "null"] }
        maxCurrentA: { type: ["number", "null"] }
        status: { type: string, enum: [available, busy, maintenance] }
        notes: { type: ["string", "null"] }
    ChargerUnitUpdate:
      type: object
      properties:
        model: { type: string }
        manufacturer: { type: string }
        serialNumber: { type: string }
        outputVoltageV: { type: ["number", "null"] }
        maxCurrentA: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    ChargerUnitStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [available, busy, maintenance] }
...
    QcApproval:
      type: object
      required: [id, productionOrderId, decision, inspectorName, inspectorRole, approvedAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        productionOrderId: { type: string, format: uuid }
        decision: { type: string, enum: [approved, rejected] }
        inspectorName: { type: string }
        inspectorRole: { type: string }
        digitalSignature: { type: ["string", "null"] }
        remarks: { type: ["string", "null"] }
        approvedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        reworkTicketId: { type: ["string", "null"], format: uuid }
    QcApprovalInput:
      type: object
      required: [decision, inspectorName]
      properties:
        decision: { type: string, enum: [approved, rejected] }
        inspectorName: { type: string }
        inspectorRole: { type: string }
        digitalSignature: { type: ["string", "null"] }
        remarks: { type: ["string", "null"] }
        failedTests: { type: array, items: { type: string } }
        failureReason: { type: ["string", "null"] }
...